# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "Termin Planungs Assistent"
  },
  "host": "localhost",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
    "/api/Appointments/{userId}": {
      "get": {
        "tags": [
          "Appointments"
        ],
        "operationId": "GetAppointments",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "uniqueItems": false,
              "type": "array",
              "items": {
                "$ref": "#/definitions/AppointmentDto"
              }
            }
          }
        }
      }
    },
    "/api/Appointments": {
      "post": {
        "tags": [
          "Appointments"
        ],
        "operationId": "Post",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "appointment",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/AppointmentDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Appointments/{id}": {
      "put": {
        "tags": [
          "Appointments"
        ],
        "operationId": "Put",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "appointment",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/AppointmentDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Appointments"
        ],
        "operationId": "Delete",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "Get",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "operationId": "Update",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "userDto",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "operationId": "Delete",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/authenticate": {
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "Authenticate",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "userDto",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/register": {
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "Register",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "userDto",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "AppointmentDto": {
      "required": [
        "entryName",
        "dateCreated",
        "status",
        "userId"
      ],
      "type": "object",
      "properties": {
        "entryName": {
          "type": "string"
        },
        "entryText": {
          "type": "string"
        },
        "dateDue": {
          "format": "date-time",
          "type": "string"
        },
        "dateCreated": {
          "format": "date-time",
          "type": "string"
        },
        "duration": {
          "format": "int32",
          "type": "integer"
        },
        "status": {
          "type": "boolean"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "id": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "UserDto": {
      "required": [
        "eMail",
        "password",
        "active"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "eMail": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "active": {
          "type": "boolean"
        }
      }
    }
  }
}